signup fixes
