+++ .gitignore
@@ -13,3 +13,6 @@ build
 .DS_Store
 .env
 npm-debug.log
+
+#yarn
+*yarn*
