+++ react-ui/src/index.js
@@ -8,7 +8,9 @@ import App from './App';
 import SignIn from './components/SignIn';
 import Users from './components/users';
 import SignOut from './components/signout';
-import SignUp from './components/NewUserContainer';
+/* eslint-disable camelcase */
+import Sign_Up from './components/NewUserContainer';
+import SignUp from './components/signup';
 import RequireAuth from './components/HOC/RequireAuth';
 
 import reducers from './reducers';
@@ -27,6 +29,7 @@ ReactDOM.render(
         <Route path="/users" component={RequireAuth(Users)} />
         <Route path="/signout" component={SignOut} />
         <Route path="/signup" component={SignUp} />
+        <Route path="/sign-up" component={Sign_Up} />
       </div>
     </Router>
   </Provider>,
