+++ server/index.js
@@ -30,13 +30,14 @@ if (cluster.isMaster) {
   app.use(bodyParser.json());
   app.use(cors());
 
+  app.use('/api/admin', adminEndpoints);
+  app.use('/api/users', userEndpoints);
   // Answer API requests.
   app.get('/api', function (req, res) {
     res.set('Content-Type', 'application/json');
     res.send('{"message":"Hello from the custom server!"}');
   });
-  app.use('/api/admin', adminEndpoints);
-  app.use('/api/users', userEndpoints);
+
 
   if (process.env.NODE_ENV === 'production') {
     // Priority serve any static files.
